requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
python-multipart==0.0.12
python-docx==1.1.2
Pillow==10.4.0
anthropic==0.34.0
python-dotenv==1.0.1
aiofiles==24.1.0
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
